remote_device/aggregator/requirements.txt
paho-mqtt
psycopg2-binary
numpy